+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   * Activity class - badges now returns an array of Osm::Activity::Badge objects
   * Activity class - files now returns an array of Osm::Activity::File objects
   * Activity class - versions now returns an array of Osm::Activity::Version objects
+  * Section class - rename extra\_records to flexi\_records
+  * Section class - flexi_records now returns an array of Osm::Section::FlexiRecord objects
 
 ## Version 0.0.11